+++ home/whatsapp.py
@@ -3,16 +3,19 @@ import logging
 import mimetypes
 from collections.abc import Iterable
 from enum import Enum
-from typing import Any
+from typing import TYPE_CHECKING, Any
 from urllib.parse import urljoin
 
 import pyparsing as pp
 import requests
 from django.conf import settings  # type: ignore
 from wagtail.images.models import Image  # type: ignore
-from wagtail.models import Locale  # type: ignore
+from wagtail.models import Locale, Revision  # type: ignore
 
 from .constants import WHATSAPP_LANGUAGE_MAPPING
+
+if TYPE_CHECKING:
+    from .models import WhatsAppTemplate
 
 logger = logging.getLogger(__name__)
 
@@ -344,6 +347,42 @@ def create_standalone_template_header_components(
         }
     )
     return components
+
+
+def submit_to_meta_action(template: "WhatsAppTemplate | Revision") -> None:
+    is_revision = isinstance(template, Revision)
+    if is_revision:
+        template = template.as_object()
+
+    template_name = template.create_whatsapp_template_name()
+    try:
+        response_json = create_standalone_whatsapp_template(
+            name=template_name,
+            message=template.message,
+            category=template.category,
+            locale=template.locale,
+            quick_replies=[b["value"]["title"] for b in template.buttons.raw_data],
+            image_obj=template.image,
+            example_values=[v["value"] for v in template.example_values.raw_data],
+        )
+        template.submission_name = template_name
+        template.submission_status = template.SubmissionStatus.SUBMITTED
+        template.submission_result = f"Success! Template ID = {response_json['id']}"
+    except TemplateSubmissionServerException as tsse:
+        logger.exception(f"TemplateSubmissionServerException: {str(tsse)} ")
+        template.submission_name = template_name
+        template.submission_status = template.SubmissionStatus.FAILED
+        template.submission_result = "An Internal Server Error has occurred.  Please try again later or contact developer support"
+    except TemplateSubmissionClientException as tsce:
+        template.submission_name = template_name
+        template.submission_status = template.SubmissionStatus.FAILED
+        template.submission_result = str(tsce)
+
+    # if we give a revision, save a new revision, otherwise overwrite the current live template
+    if is_revision:
+        template.save_revision()
+    else:
+        template.save()
 
 
 def create_standalone_whatsapp_template(